chore: move `useStore` to subfolder
